Fixed Bugs

+++ cpp-seq-codes/para.hpp
@@ -5,9 +5,9 @@ int M = 100 ;
 int max_steps = 100000 ; 
 
 std::vector<int> E(max_steps + 1, 0) ; 
-std::vector<int> m(max_steps + 1, 0) ;
+std::vector<float> m(max_steps + 1, 0) ;
 
 int T = 10 ; 
-int eps = 0.01 ; 
+float eps = 0.01 ;
